+++ Rakefile
@@ -33,6 +33,7 @@ spec = Hoe.spec 'mime-types' do
   extra_dev_deps << ['minitest-autotest', '~> 1.0']
   extra_dev_deps << ['minitest-focus', '~> 1.0']
   extra_dev_deps << ['minitest-bonus-assertions', '~> 2.0']
+  extra_dev_deps << ['minitest-hooks', '~> 1.4']
   extra_dev_deps << ['rake', '~> 10.0']
   extra_dev_deps << ['fivemat', '~> 1.3' ]
   extra_dev_deps << ['minitest-rg', '~> 5.2']
