+++ test/minitest_helper.rb
@@ -8,5 +8,6 @@ require 'fivemat/minitest/autorun'
 require 'minitest/focus'
 require 'minitest/rg'
 require 'minitest-bonus-assertions'
+require 'minitest/hooks'
 
 ENV['RUBY_MIME_TYPES_LAZY_LOAD'] = 'yes'
